+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to semantic versioning for configuration changes.
 
 ## [Unreleased]
 
+### Auto-updated - 2026-01-05 08:00:11
+- Added: scripts/net2_user_activity_daemon.py
+- Added: services/net2-daemon.service
+
+## [Unreleased]
+
 ### Auto-updated - 2026-01-05 07:42:48
 - Deleted: test_commit.txt